pushing for partner

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -3,8 +3,11 @@ package com.techelevator.tenmo.dao;
 
 import com.techelevator.tenmo.model.Account;
 
+import java.math.BigDecimal;
+
 public interface AccountDao {
 
     Account getAccountByUserId(int user_id);
-    Account moveMoney(double amountToTransfer, int withdraw_id, int deposit_id);
+
+    Account moveMoney(BigDecimal amountToTransfer, int withdraw_id, int deposit_id);
 }
